usersync/src/githubGetCommitAuthors_test.go
package usersync

import (
	"fmt"
	"testing"

	"github.com/OpenQDev/GoGitguru/util/setup"
	"github.com/OpenQDev/GoGitguru/util/testhelpers"
)

func TestGithubGetCommitAuthors(t *testing.T) {
	env := setup.ExtractAndVerifyEnvironment("../.env")
	ghAccessToken := env.GhAccessToken
	targetLiveGithub := env.TargetLiveGithub

	var serverUrl string
	if targetLiveGithub {
		serverUrl = "https://api.github.com/graphql"
	} else {
		serverUrl = "https://api.github.com/graphql"
	}

	tests := GithubGetCommitAuthorsTestCases()

	for _, tt := range tests {
		testhelpers.CheckTestSkip(t, testhelpers.Targets(
			testhelpers.RUN_ALL_TESTS,
		), tt.name)

		t.Run(tt.name, func(t *testing.T) {
			var commitDetails string

			for i, commit := range tt.commits {
				commitDetails += fmt.Sprintf(`commit_%d: object(oid: "%s") {
					...commitDetails
				}
				`, i, commit)
			}

			query := fmt.Sprintf(`{
				rateLimit {
					limit
					used
					resetAt
				}
				repository(owner: "%s", name: "%s") {
					%s
				}
			}
			`, tt.owner, tt.repo, commitDetails) + AUTHOR_GRAPHQL_FRAGMENT

			fmt.Println(ghAccessToken)
			resp, err := GithubGetCommitAuthors(query, ghAccessToken, serverUrl)

			if (err != nil) != tt.wantErr {
				t.Errorf("GithubGetCommitAuthors() error = %v, wantErr %v", err, tt.wantErr)
			}

			actualRestId := resp.Data.Repository["commit_1"].Author.User.GithubRestID
			expectedRestId := 93455288
			if actualRestId != expectedRestId {
				t.Errorf("GithubGetCommitAuthors() unexpected return. expect rest ID of %d but got %d", expectedRestId, actualRestId)
			}
		})
	}
}
